refactor: extract Provider.Profiles
Move provider-profile create/draft/update/complete, the profile reads and
form changesets, and admin verify/unverify onto Provider.Profiles, along with
the insert/persist helpers, publish_verification_event, and the two @profile_*
attributes. verify/unverify live here (not Verification) because they mutate
ProviderProfile. Facade delegates.
